feat(simulator)!: removed simulator feature
- moved simulator from mousefood/examples to examples/simulator
- removed `simulator` feature
- ci: run `fmt` and `clippy` on all workspace crates
