+++ mousefood/Cargo.toml
@@ -16,7 +16,6 @@ exclude.workspace = true
 ratatui-core.workspace = true
 thiserror.workspace = true
 embedded-graphics.workspace = true
-embedded-graphics-simulator = { workspace = true, optional = true }
 embedded-graphics-unicodefonts = { workspace = true, optional = true }
 weact-studio-epd = { workspace = true, optional = true }
 
@@ -28,13 +27,8 @@ paste.workspace = true
 [features]
 default = ["fonts"]
 std = ["thiserror/std", "ratatui-core/std"]
-simulator = ["dep:embedded-graphics-simulator"]
 fonts = ["dep:embedded-graphics-unicodefonts"]
 epd-weact = ["dep:weact-studio-epd"]
-
-[[example]]
-name = "simulator"
-required-features = ["simulator"]
 
 [lints]
 workspace = true
