--- mousefood/examples/simulator.rs
@@ -1,27 +1,0 @@
-//! Example using a simulator.
-use mousefood::embedded_graphics::geometry;
-use mousefood::error::Error;
-use mousefood::prelude::*;
-use mousefood::simulator::SimulatorDisplay;
-use ratatui::widgets::{Block, Paragraph, Wrap};
-use ratatui::{Frame, Terminal, style::*};
-
-fn main() -> Result<(), Error> {
-    let mut display = SimulatorDisplay::<Bgr565>::new(geometry::Size::new(128, 64));
-    let backend: EmbeddedBackend<SimulatorDisplay<_>, _> =
-        EmbeddedBackend::new(&mut display, EmbeddedBackendConfig::default());
-    let mut terminal = Terminal::new(backend)?;
-
-    loop {
-        terminal.draw(draw)?;
-    }
-}
-
-fn draw(frame: &mut Frame) {
-    let text = "Ratatui on embedded devices!";
-    let paragraph = Paragraph::new(text.dark_gray()).wrap(Wrap { trim: true });
-    let bordered_block = Block::bordered()
-        .border_style(Style::new().yellow())
-        .title("Mousefood");
-    frame.render_widget(paragraph.block(bordered_block), frame.area());
-}
