+++ mousefood/src/error.rs
@@ -9,8 +9,4 @@ pub enum Error {
     /// Selected [`ClearType`] is not supported by Mousefood.
     #[error("ClearType::{0} is not supported by Mousefood")]
     ClearTypeUnsupported(alloc::string::String),
-    /// Simulator window was closed.
-    #[cfg(feature = "simulator")]
-    #[error("simulator window closed")]
-    SimulatorQuit,
 }
